Create New Pixel Helpers.dm
This is a library containing some functions that make the native pixel movement in BYOND easier to work with. Along with the Pixel Positions library to get world pixel positions of atoms, you can use this library to set world pixel positions of atoms and move them around by pixels.